hw2 refactor: custom hook
