+++ homeworks/week21/hw2/game-gomoku/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["react-app"],
+  "rules": {
+    "react/prop-types": "warn"
+  }
+}
