--- homeworks/week21/hw2/game-gomoku/src/App.test.js
@@ -1,8 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import Gomoku from './Gomoku';
-
-test('renders learn react link', () => {
-  render(<Gomoku />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
